Namespace updates
