+++ src/workstation-gui/app/controllers/api_v1_controller.rb
@@ -14,31 +14,20 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-class AuthenticationController < ApplicationController
-  before_action :validate_creds, only: %i[login]
-  skip_before_action :authenticate_api_requests!, only: %i[login]
+class ApiV1Controller < ApplicationController
+  protect_from_forgery with: :null_session
+  before_action :authenticate_api_requests!
 
-  def login
+  def authenticate_api_requests!
+    header = request.headers["Authorization"]
+    header = header.split(" ").last if header
+
+    decoded = Auth::JwtToken.decode(header)
     access_key = Rails.application.credentials.access_key.to_s
-
-    if params[:access_key] == access_key
-      token = Auth::JwtToken.encode(access_key: params[:access_key])
-
-      render json: {
-        token: token
-      }, status: :ok
-    else
-      render json: { errors: "Invalid credentials" }, status: :unauthorized
+    if decoded["access_key"] != access_key
+      render json: { message: "Invalid auth credentials", status: 401 }, status: :unauthorized
     end
-  end
-
-  def testing
-    render json: {status: "okay", message: "success" }
-  end
-
-  private
-
-  def validate_creds
-    render json: { errors: "Access key is required"} unless params.key?(:access_key)
+  rescue JWT::DecodeError => e
+    render json: { message: "Unable to process the auth token: #{e.message}", status: 401 }, status: :unauthorized
   end
 end
